Try to fix circular dependency in tests 2
