+++ code/go/0chain.net/smartcontract/common/wallet.go
@@ -1,8 +1,11 @@
 package common
 
 import (
+	"encoding/hex"
+
 	"0chain.net/chaincore/config"
 	"0chain.net/core/common"
+	"0chain.net/core/encryption"
 	"github.com/0chain/common/core/logging"
 )
 
@@ -12,7 +15,7 @@ func ValidateWallet(publicKey, delegateWalletID string) *common.Error {
 		return nil
 	}
 
-	operationalWalletID, err := common.GetWalletIdFromPublicKey(publicKey)
+	operationalWalletID, err := GetWalletIdFromPublicKey(publicKey)
 	if err != nil {
 		return common.NewError("add_sharder",
 			"couldn't decode publick key to compare to delegate wallet")
@@ -25,4 +28,14 @@ func ValidateWallet(publicKey, delegateWalletID string) *common.Error {
 	}
 
 	return nil
+}
+
+/*GetWalletIdFromPublicKey - given the PK of the provider, return the its operational walletID i.e. the key used to sign the txns */
+func GetWalletIdFromPublicKey(pk string) (string, error) {
+	publicKeyBytes, err := hex.DecodeString(pk)
+	if err != nil {
+		return "", err
+	}
+	operationalClientID := encryption.Hash(publicKeyBytes)
+	return operationalClientID, nil
 }
